# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/central/states": {
      "get": {
        "operationId": "CentralController_listStates",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all State Governments with metrics",
        "tags": [
          "central"
        ]
      },
      "post": {
        "operationId": "CentralController_createState",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new State Government (1 State Admin enforced)",
        "tags": [
          "central"
        ]
      }
    },
    "/api/v1/central/states/{id}": {
      "get": {
        "operationId": "CentralController_getState",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get State Government by ID",
        "tags": [
          "central"
        ]
      },
      "patch": {
        "operationId": "CentralController_updateState",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update State Government metadata",
        "tags": [
          "central"
        ]
      },
      "delete": {
        "operationId": "CentralController_deleteState",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Deactivate State Government",
        "tags": [
          "central"
        ]
      }
    },
    "/api/v1/central/states/{id}/details": {
      "get": {
        "operationId": "CentralController_getStateDetails",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get full State Government monitoring details",
        "tags": [
          "central"
        ]
      }
    },
    "/api/v1/central/states/{id}/status": {
      "patch": {
        "operationId": "CentralController_setStateStatus",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Activate or Deactivate State Government",
        "tags": [
          "central"
        ]
      }
    },
    "/api/v1/central/states/{id}/admin/status": {
      "patch": {
        "operationId": "CentralController_setStateAdminStatus",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Activate or Deactivate State Admin Account",
        "tags": [
          "central"
        ]
      }
    },
    "/api/v1/central/states/{id}/admin/reset-password": {
      "post": {
        "operationId": "CentralController_resetStateAdminPassword",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reset State Admin temporary credentials",
        "tags": [
          "central"
        ]
      }
    },
    "/api/v1/central/analytics/revenue": {
      "get": {
        "operationId": "CentralController_getStateWiseRevenue",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Must be CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "State-wise revenue analysis",
        "tags": [
          "central"
        ]
      }
    },
...
  },
  "info": {
    "title": "UCSDP Backend API",
    "description": "Unified Citizen Service and Delivery Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateStateDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}